requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
psycopg2==2.9.1
pytz==2021.1
